@@ -31,10 +37,11 @@
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <maven.compiler.target>1.8</maven.compiler.target>
-        <flink.version>1.3.2</flink.version>
-        <jackson.version>2.4.0</jackson.version>
+        <java.version>11</java.version>
+        <maven.compiler.source>${java.version}</maven.compiler.source>
+        <maven.compiler.target>${java.version}</maven.compiler.target>
+        <flink.version>1.11.2</flink.version>
+        <jackson.version>2.11.0</jackson.version>
     </properties>
     <distributionManagement>
         <snapshotRepository>
@@ -104,13 +120,22 @@
                             <goal>jar</goal>
                         </goals>
                         <configuration>
-                            <additionalparam>-Xdoclint:none</additionalparam>
+                            <doclint>none</doclint>
+                            <javadocExecutable>${java.home}/bin/javadoc</javadocExecutable>
+                            <sourcepath>src/main/java</sourcepath>
+                            <!--
+                            We ran into issues with Java 11 (named vs unnamed modules)
+                            Checkout the solution here: https://stackoverflow.com/a/63475569
+                            -->
+                            <source>1.8</source>
+                            <detectJavaApiLink>false</detectJavaApiLink>
                         </configuration>
                     </execution>
                 </executions>
             </plugin>
             <plugin>
                 <artifactId>maven-assembly-plugin</artifactId>
+                <version>3.3.0</version>
                 <configuration>
                     <archive>
                         <manifest>
