@@ -2,8 +2,8 @@ package com.formulasearchengine.mathosphere.mlp;
 
 import com.formulasearchengine.mathosphere.mlp.cli.MachineLearningDefinienExtractionConfig;
 import com.formulasearchengine.mathosphere.mlp.contracts.JsonSerializerMapper;
-import com.formulasearchengine.mathosphere.mlp.contracts.TextAnnotatorMapper;
-import com.formulasearchengine.mathosphere.mlp.contracts.TextExtractorMapper;
+import com.formulasearchengine.mathosphere.mlp.contracts.WikiTextAnnotatorMapper;
+import com.formulasearchengine.mathosphere.mlp.contracts.WikiTextPageExtractorMapper;
 import com.formulasearchengine.mathosphere.mlp.pojos.EvaluationResult;
 import com.formulasearchengine.mathosphere.mlp.ml.WekaLearner;
 import com.formulasearchengine.mathosphere.mlp.pojos.ParsedWikiDocument;
@@ -32,8 +32,8 @@ public class MachineLearningModelGenerator {
       ExecutionEnvironment env = ExecutionEnvironment.getExecutionEnvironment();
       env.setParallelism(config.getParallelism());
       DataSource<String> source = FlinkMlpRelationFinder.readWikiDump(config, env);
-      DataSet<ParsedWikiDocument> documents = source.flatMap(new TextExtractorMapper())
-        .map(new TextAnnotatorMapper(config));
+      DataSet<ParsedWikiDocument> documents = source.flatMap(new WikiTextPageExtractorMapper())
+        .map(new WikiTextAnnotatorMapper(config));
       ArrayList<GoldEntry> gold = (new Evaluator()).readGoldEntries(new File(config.getGoldFile()));
       DataSet<WikiDocumentOutput> instances = documents.map(new SimpleFeatureExtractorMapper(config, gold));
       //process parsed wikipedia
