@@ -6,7 +6,7 @@ import com.formulasearchengine.mathosphere.mlp.cli.CliParams;
 import com.formulasearchengine.mathosphere.mlp.cli.ListCommandConfig;
 import com.formulasearchengine.mathosphere.mlp.cli.MlpCommandConfig;
 import com.formulasearchengine.mathosphere.mlp.contracts.CreateCandidatesMapper;
-import com.formulasearchengine.mathosphere.mlp.contracts.TextAnnotatorMapper;
+import com.formulasearchengine.mathosphere.mlp.contracts.WikiTextAnnotatorMapper;
 import com.formulasearchengine.mathosphere.mlp.pojos.ParsedWikiDocument;
 import com.formulasearchengine.mathosphere.mlp.pojos.RawWikiDocument;
 import com.formulasearchengine.mathosphere.mlp.pojos.Relation;
@@ -22,6 +22,15 @@ import java.io.FileNotFoundException;
 import java.io.PrintWriter;
 import java.util.List;
 
+/**
+ * The main Math Language Processor (MLP) class following the original idea of the MLP.
+ * That means, it analyzes wikitext, extracts all mathematical expressions and nouns (as definiens).
+ * It creates candidates of identifier-definiens via {@link CreateCandidatesMapper} and scores them
+ * based on distance calculations.
+ *
+ * The main entry point is {@link #run(MlpCommandConfig)}.
+ * The method {@link #list(ListCommandConfig)} is for testing.
+ */
 public class RelationExtractor {
 
   public static void main(String[] args) throws Exception {
@@ -37,7 +46,7 @@ public class RelationExtractor {
       List<Relation> relations = output.getRelations();
       CSVPrinter printer = CSVFormat.DEFAULT.withRecordSeparator("\n").print(pw);
       for (Relation r : relations) {
-        String[] record = {r.getIdentifier(), r.getDefinition(), Double.toString(r.getScore())};
+        String[] record = {r.getMathTag().getContent(), r.getDefinition(), Double.toString(r.getScore())};
         printer.printRecord(record);
       }
       printer.flush();
@@ -45,15 +54,28 @@ public class RelationExtractor {
     }
   }
 
+  /**
+   * The main pipeline. It creates {@link WikiDocumentOutput} for a given input. The output
+   * already contains all information (scored identifier-definiens candidates).
+   * @param config configuration
+   * @return final document with scored identifier-definiens pairs.
+   * @throws Exception if something went wrong
+   */
   private static WikiDocumentOutput getWikiDocumentOutput(MlpCommandConfig config) throws Exception {
-    TextAnnotatorMapper annotator = new TextAnnotatorMapper(config);
-    annotator.open(null);
-
     String filePath = config.getInput();
     String text = FileUtils.readFileToString(new File(filePath), "UTF-8");
-    RawWikiDocument doc = new RawWikiDocument(filePath, 0, text);
+
+    // create document based on given input
+    RawWikiDocument doc = new RawWikiDocument(text);
+
+    // load wikitext annotator
+    WikiTextAnnotatorMapper annotator = new WikiTextAnnotatorMapper(config);
+    annotator.open(null);
+
+    // annotate (PoS tagging)
     ParsedWikiDocument parsedDocument = annotator.map(doc);
 
+    // create candidates of identifier-definien pairs and score them
     CreateCandidatesMapper mlp = new CreateCandidatesMapper(config);
     return mlp.map(parsedDocument);
   }
