@@ -3,7 +3,7 @@ package com.formulasearchengine.mathosphere.mlp.contracts;
 import com.formulasearchengine.mathosphere.mlp.cli.TagsCommandConfig;
 import com.formulasearchengine.mathosphere.mlp.pojos.MathTag;
 import com.formulasearchengine.mathosphere.mlp.pojos.RawWikiDocument;
-import com.formulasearchengine.mathosphere.mlp.text.MathConverter;
+import com.formulasearchengine.mathosphere.mlp.text.WikiTextParser;
 import org.apache.flink.api.common.functions.FlatMapFunction;
 import org.apache.flink.util.Collector;
 
@@ -20,10 +20,10 @@ public class TagExtractionMapper implements FlatMapFunction<RawWikiDocument, Mat
 
     @Override
     public void flatMap(RawWikiDocument rawWikiDocument, Collector<MathTag> collector) throws Exception {
-        final MathConverter converter = new MathConverter(rawWikiDocument.text, rawWikiDocument.title, config);
+        final WikiTextParser converter = new WikiTextParser(rawWikiDocument, config);
         converter.setSkipHiddenMath(true);
         converter.processTags();
-        for (MathTag tag : converter.getMathTags()) {
+        for (MathTag tag : converter.getMetaLibrary().getFormulaLib().values()) {
             collector.collect(tag);
         }
     }
