@@ -1,91 +1,93 @@
 package com.formulasearchengine.mathosphere.mlp.pojos;
 
 import com.google.common.collect.Multiset;
+import edu.stanford.nlp.trees.GrammaticalStructure;
 
+import java.util.Collection;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 public class ParsedWikiDocument {
-
-  private List<WikidataLink> links;
   private String title;
-  private Multiset<String> identifiers;
-  private List<MathTag> formulas;
   private List<Sentence> sentences;
-  private Map<String, WikidataLink> linkMap = null;
+  private DocumentMetaLib lib;
+  private Multiset<String> identifier;
 
   public ParsedWikiDocument() {
   }
 
-  public ParsedWikiDocument(String title, Multiset<String> identifiers, List<MathTag> formulas,
-                            List<Sentence> sentences) {
+  public ParsedWikiDocument(String title, List<Sentence> sentences, DocumentMetaLib lib) {
     this.title = title;
-    this.identifiers = identifiers;
-    this.formulas = formulas;
+    this.lib = lib;
     this.sentences = sentences;
-  }
-
-  public ParsedWikiDocument(String title, Multiset<String> identifiers, List<MathTag> formulas, List<Sentence> sentences, List<WikidataLink> links) {
-    this.title = title;
-    this.identifiers = identifiers;
-    this.formulas = formulas;
-    this.sentences = sentences;
-    this.links = links;
-  }
-
-
-  public String getTitle() {
-    return title;
   }
 
   public void setTitle(String title) {
     this.title = title;
   }
 
-  public List<Sentence> getSentences() {
-    return sentences;
-  }
-
-  public Multiset<String> getIdentifiers() {
-    return identifiers;
-  }
-
-  public void setIdentifiers(Multiset<String> identifiers) {
-    this.identifiers = identifiers;
-  }
-
-  public List<MathTag> getFormulas() {
-    return formulas;
-  }
-
-  public void setFormulas(List<MathTag> formulas) {
-    this.formulas = formulas;
-  }
-
   public void setSentences(List<Sentence> sentences) {
     this.sentences = sentences;
   }
 
-  public List<WikidataLink> getLinks() {
-    return links;
+  public void setLib(DocumentMetaLib lib) {
+    this.lib = lib;
   }
 
-  public ParsedWikiDocument setLinks(List<WikidataLink> links) {
-    this.links = links;
-    this.linkMap = null;
-    return this;
+  public void setIdentifier(Multiset<String> identifier) {
+    this.identifier = identifier;
   }
 
-  public Map<String, WikidataLink> getLinkMap() {
-    if (linkMap == null) {
-      linkMap = new HashMap<>();
-      if (links != null) {
-        for (WikidataLink i : links) {
-          linkMap.put(i.getContentHash(), i);
-        }
-      }
-    }
-    return linkMap;
+  public String getTitle() {
+    return title;
+  }
+
+  public List<Sentence> getSentences() {
+    return sentences;
+  }
+
+  public DocumentMetaLib getLib() {
+    return lib;
+  }
+
+  public MathTagGraph getFormulaGraph() {
+    return lib.getGraph();
+  }
+
+  public Collection<MathTag> getFormulae() {
+    return lib.getFormulaLib().values();
+  }
+
+  public Map<String, MathTag> getFormulaeMap() {
+    return lib.getFormulaLib();
+  }
+
+  public Map<String, SpecialToken> getCitationMap() {
+    return lib.getCiteLib();
+  }
+
+  public Map<String, SpecialToken> getLinkMap() {
+    return lib.getLinkLib();
+  }
+
+  public Multiset<String> getIdentifier() {
+    return identifier;
+  }
+
+  /**
+   * @deprecated we switching to MOI instead of single identifiers.
+   */
+  @Deprecated
+  public void setIdentifiers(Multiset<String> identifier) {
+    this.identifier = identifier;
+  }
+
+  /**
+   * @deprecated we switching to MOI instead of single identifiers.
+   */
+  @Deprecated
+  public Multiset<String> getIdentifiers() {
+    return this.identifier;
   }
 }
