@@ -44,7 +44,8 @@ public class SimpleFeatureExtractorMapper implements MapFunction<ParsedWikiDocum
     Multiset<String> frequencies = aggregateWords(sentences);
 
     int maxFrequency = getMaxFrequency(frequencies, doc.getTitle());
-    if (extractionForTraining) {
+    GoldEntry goldEntry = getGoldEntryByTitle(goldEntries, doc.getTitle());
+    if (extractionForTraining && goldEntry != null) {
       return getIdentifiersWithGoldInfo(doc, allIdentifierDefininesCandidates, sentences, identifierSentenceDistanceMap, frequencies, maxFrequency);
     } else {
       return getAllIdentifiers(doc, allIdentifierDefininesCandidates, sentences, identifierSentenceDistanceMap, frequencies, maxFrequency);
@@ -54,7 +55,7 @@ public class SimpleFeatureExtractorMapper implements MapFunction<ParsedWikiDocum
   private WikiDocumentOutput getIdentifiersWithGoldInfo(ParsedWikiDocument doc, List<Relation> allIdentifierDefininesCandidates, List<Sentence> sentences, Map<String, Integer> identifierSentenceDistanceMap, Multiset<String> frequencies, double maxFrequency) {
     GoldEntry goldEntry = getGoldEntryByTitle(goldEntries, doc.getTitle());
     final Integer fid = Integer.parseInt(goldEntry.getFid());
-    Stream<MathTag> stream = doc.getFormulas().stream();
+    Stream<MathTag> stream = doc.getFormulaeMap().values().stream();
     stream = stream.filter( e -> e.getMarkUpType().equals(WikiTextUtils.MathMarkUpType.LATEX) );
     List<MathTag> list = stream.collect( Collectors.toList() );
     MathTag seed = null;
@@ -95,7 +96,7 @@ public class SimpleFeatureExtractorMapper implements MapFunction<ParsedWikiDocum
     }
     LOGGER.info("extracted {} relations from {}", allIdentifierDefininesCandidates.size(), doc.getTitle());
     WikiDocumentOutput result = new WikiDocumentOutput(doc.getTitle(), goldEntry.getqID(), allIdentifierDefininesCandidates, null);
-    result.setMaxSentenceLength(doc.getSentences().stream().map(s -> s.getWords().size()).max(Comparator.naturalOrder()).get());
+    result.setMaxSentenceLength(doc.getSentences().stream().map(s -> s.getWords().size()).max(Comparator.naturalOrder()).orElse(0));
     return result;
   }
 
@@ -134,7 +135,7 @@ public class SimpleFeatureExtractorMapper implements MapFunction<ParsedWikiDocum
     for (String identifier : identifiers) {
       for (int i = 0; i < sentences.size(); i++) {
         Sentence sentence = sentences.get(i);
-        if (sentence.contains(identifier)) {
+        if (sentence.containsIdentifier(identifier)) {
           result.put(identifier, i);
           break;
         }
