@@ -1,5 +1,8 @@
 package com.formulasearchengine.mathosphere.mlp.text;
 
+import com.fasterxml.jackson.core.JsonFactory;
+import com.fasterxml.jackson.core.JsonParser;
+import com.fasterxml.jackson.core.JsonToken;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.guava.GuavaModule;
 
@@ -13,9 +16,6 @@ import com.formulasearchengine.mathosphere.mlp.cli.CountCommandConfig;
 import org.apache.commons.csv.CSVFormat;
 import org.apache.commons.csv.CSVPrinter;
 import org.apache.flink.api.java.tuple.Tuple2;
-import org.codehaus.jackson.JsonFactory;
-import org.codehaus.jackson.JsonParser;
-import org.codehaus.jackson.JsonToken;
 
 import java.io.FileInputStream;
 import java.io.IOException;
