@@ -13,27 +13,35 @@ import org.junit.Assert;
 import org.junit.Test;
 
 import static com.formulasearchengine.mathosphere.mlp.text.WikiTextUtilsTest.getTestResource;
-import static junit.framework.Assert.assertEquals;
-import static junit.framework.Assert.assertTrue;
+import static junit.framework.Assert.*;
 import static org.hamcrest.CoreMatchers.containsString;
 
 /**
  * Created by Moritz on 30.09.2015.
+ * Fixed by Andre
  */
 public class MathTagTest {
 
     @Test
-    public void testGetContentHash() throws Exception {
-        MathTag t = new MathTag(1, "<math><mrow><mo>x</mo><mrow></math>", WikiTextUtils.MathMarkUpType.MATHML);
-        assertEquals("2cf05995ef521b456aa419201d160406", t.getContentHash());
+    public void testHash() {
+        String content = "<math><mrow><mo>x</mo><mrow></math>";
+        MathTag t1 = new MathTag(content, WikiTextUtils.MathMarkUpType.MATHML);
+        t1.addPosition(new Position(2,1,1));
+        MathTag t2 = new MathTag(content, WikiTextUtils.MathMarkUpType.LATEX);
+        t2.addPosition(new Position(3));
+        t2.addPosition(new Position(4,1));
+
+        assertEquals("The hash should be only generated based on the content.", t1.getContentHash(), t2.getContentHash());
+        assertNotSame("The real hash should be different", t1.hashCode(), t2.hashCode());
+        assertEquals("The placeholder should be only generated based on the content.", t1.placeholder(), t2.placeholder());
     }
 
     @Test
     public void testGetIdentifier() throws Exception {
-        MathTag tagX = new MathTag(1, "<math xmlns=\"http://www.w3.org/1998/Math/MathML\"><mrow><mi>x</mi></mrow></math>", WikiTextUtils.MathMarkUpType.MATHML);
+        MathTag tagX = new MathTag("<math xmlns=\"http://www.w3.org/1998/Math/MathML\"><mrow><mi>x</mi></mrow></math>", WikiTextUtils.MathMarkUpType.MATHML);
         assertEquals(ImmutableSet.of("x"), tagX.getIdentifier(true, true).elementSet());
         assertEquals(ImmutableSet.of("x"), tagX.getIdentifier(false, true).elementSet());
-        MathTag schrödinger = new MathTag(1, getTestResource("com/formulasearchengine/mathosphere/mlp/schrödinger_eq.xml"), WikiTextUtils.MathMarkUpType.MATHML);
+        MathTag schrödinger = new MathTag(getTestResource("com/formulasearchengine/mathosphere/mlp/schrödinger_eq.xml"), WikiTextUtils.MathMarkUpType.MATHML);
         //MathTag schrödingerTex = new MathTag(1,"i\\hbar\\frac{\\partial}{\\partial t}\\Psi(\\mathbb{r},\\,t)=-\\frac{\\hbar^{2}}{2m}" +
         //  "\\nabla^{2}\\Psi(\\mathbb{r},\\,t)+V(\\mathbb{r})\\Psi(\\mathbb{r},\\,t).", WikiTextUtils.MathMarkUpType.LATEX);
         ImmutableMultiset<String> lIds = ImmutableMultiset.of("i",
@@ -48,7 +56,7 @@ public class MathTagTest {
 
     @Test
     public void testGetJson() throws JsonProcessingException {
-        MathTag tag = new MathTag(1, "a+b", WikiTextUtils.MathMarkUpType.LATEX);
+        MathTag tag = new MathTag("a+b", WikiTextUtils.MathMarkUpType.LATEX);
         Assert.assertThat(tag.toJson(), containsString("\"inputhash\""));
         Assert.assertThat(tag.toJson(), containsString("\"input\""));
         Assert.assertThat(tag.toJson(), containsString("\"type\""));
@@ -56,7 +64,7 @@ public class MathTagTest {
 
     @Test
     public void testGetJsonValidation() throws Exception {
-        MathTag tag = new MathTag(1, "a+b", WikiTextUtils.MathMarkUpType.LATEX);
+        MathTag tag = new MathTag("a+b", WikiTextUtils.MathMarkUpType.LATEX);
         final String real = "[" + tag.toJson() + "]";
         final JsonNode jsonSchema = JsonLoader.fromResource("/formula_list_schema.json");
         final JsonSchemaFactory factory = JsonSchemaFactory.byDefault();
