@@ -2,6 +2,7 @@ package com.formulasearchengine.mathosphere.mlp.text;
 
 import com.formulasearchengine.mathosphere.mlp.PatternMatchingRelationFinder;
 import com.formulasearchengine.mathosphere.mlp.cli.FlinkMlpCommandConfig;
+import com.formulasearchengine.mathosphere.mlp.pojos.DocumentMetaLib;
 import com.formulasearchengine.mathosphere.mlp.pojos.MathTag;
 import com.formulasearchengine.mathosphere.mlp.pojos.Sentence;
 import com.formulasearchengine.mathosphere.mlp.pojos.Word;
@@ -14,7 +15,9 @@ import org.apache.logging.log4j.LogManager;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Arrays;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import static org.junit.Assert.assertEquals;
 
@@ -33,21 +36,14 @@ public class PosTaggerTestGer {
   public void simpleGermanTest() throws Exception {
     FlinkMlpCommandConfig cfg = FlinkMlpCommandConfig.test();
     cfg.setModel(GER);
-    PosTagger nlpProcessor = PosTagger.create(cfg);
+    TextAnnotator annotator = new TextAnnotator(cfg);
     String text = "Dies ist ein simpler Beispieltext.";
-
-    List<MathTag> mathTags = WikiTextUtils.findMathTags(text);
-
-    String newText = WikiTextUtils.replaceAllFormulas(text, mathTags);
-    String cleanText = WikiTextUtils.extractPlainText(newText);
-
-    List<Sentence> result = nlpProcessor.process(cleanText, mathTags);
-
+    List<Sentence> result = annotator.annotate(text, new DocumentMetaLib(cfg));
     List<Word> expected = Arrays.asList(w("Dies", "PDS"), w("ist", "VAFIN"), w("ein", "ART"), w("simpler", "ADJA"), w("Beispieltext", "NN"),
         w(".", "$."));
 
     List<Word> sentence = result.get(0).getWords();
-    assertEquals(expected, sentence.subList(0, expected.size()));
+    assertEquals(expected.toString(), sentence.subList(0, expected.size()).toString());
     LOGGER.debug("full result: {}", result);
   }
 
@@ -57,16 +53,14 @@ public class PosTaggerTestGer {
 
     FlinkMlpCommandConfig cfg = FlinkMlpCommandConfig.test();
     cfg.setModel(GER);
-    PosTagger nlpProcessor = PosTagger.create(cfg);
+    TextAnnotator annotator = new TextAnnotator(cfg);
 
-
-    List<MathTag> mathTags = WikiTextUtils.findMathTags(text);
-
-    String newText = WikiTextUtils.replaceAllFormulas(text, mathTags);
     long t0 = System.nanoTime();
+    List<MathTag> mathTags = WikiTextUtils.findMathTags(text);
+    String newText = WikiTextUtils.replaceAllFormulas(text, mathTags);
     String cleanText = WikiTextUtils.extractPlainText(newText);
-    System.out.println((System.nanoTime() - t0) / 1000000 + "ms for cleaning.");
-    List<Sentence> result = nlpProcessor.process(cleanText, mathTags);
+    System.out.println((System.nanoTime() - t0) / 1000000 + "ms for processing.");
+//    List<Sentence> result = annotator.annotate(cleanText, mathTags);
   }
 
   public static Word w(String word, String tag) {
